pyproject.toml
[tool.poetry]
name = "aligned"
version = "0.0.1a"
description = "A scalable feature store that makes it easy to align offline and online ML systems"
authors = ["Mats E. Mollestad <mats@mollestad.no>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/otovo/aladdin"
repository = "https://github.com/otovo/aladdin"
keywords = [
    'python',
    'typed',
    'ml',
    'prediction',
    'feature',
    'store',
    'feature-store',
    'feast',
    'tecton',
]
classifiers = [
    'Development Status :: 3 - Alpha',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.10',
    'Topic :: Internet :: WWW/HTTP',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
    'Topic :: Software Development',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Application Frameworks',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]
packages = [
    { include = "aligned" },
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pandas = "^1.3.1"
fastapi = { version = "^0.77.1", optional = true }
uvicorn = { version = "^0.17.6", optional = true }
redis = { version = "^4.3.1", optional = true }
mashumaro = "^3.0.1"
dill = "^0.3.4"
aioaws = { version = "^0.12", optional = true }
databases = { version = "^0.5.5", optional = true }
asyncpg = { version = "^0.25.0", optional = true }
pyarrow = "^8.0.0"
Jinja2 = "^3.1.2"
nest-asyncio = "^1.5.5"
asgi-correlation-id = { version = "^3.0.0", optional = true }
dask = {version = "^2022.7.0", extras = ["dataframe"], optional = true}
pandera = {version = "^0.13.3", optional = true}

[tool.poetry.extras]
aws = ["aioaws", "databases"]
psql = ["databases", "asyncpg"]
redis = ["redis"]
server = ["asgi-correlation-id", "fastapi", "uvicorn"]
dask = ["dask"]
pandera = ["pandera"]

[tool.poetry.dev-dependencies]
types-redis = "^4.2.6"
pytest-mock = "^3.8.1"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aligned = 'aligned.cli:cli'

[tool.black]
line-length = 110
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 110
